Remove ember-test-helpers-fire-native-events feature flagging.
