+++ features.json
@@ -1,7 +1,6 @@
 {
   "features": {
     "features-stripped-test": null,
-    "ember-test-helpers-fire-native-events": true,
     "ember-routing-route-configured-query-params": null,
     "ember-libraries-isregistered": null,
     "ember-metal-ember-assign": true,
