--- packages/ember-testing/lib/events/_native.js
@@ -1,108 +1,0 @@
-import jQuery from 'ember-views/system/jquery';
-import run from 'ember-metal/run_loop';
-
-const DEFAULT_EVENT_OPTIONS = { canBubble: true, cancelable: true };
-const KEYBOARD_EVENT_TYPES = ['keydown', 'keypress', 'keyup'];
-const MOUSE_EVENT_TYPES = ['click', 'mousedown', 'mouseup', 'dblclick', 'mouseenter', 'mouseleave', 'mousemove', 'mouseout', 'mouseover'];
-
-export function focus(el) {
-  if (!el) { return; }
-  let $el = jQuery(el);
-  if ($el.is(':input, [contenteditable=true]')) {
-    let type = $el.prop('type');
-    if (type !== 'checkbox' && type !== 'radio' && type !== 'hidden') {
-      run(null, function() {
-        // Firefox does not trigger the `focusin` event if the window
-        // does not have focus. If the document doesn't have focus just
-        // use trigger('focusin') instead.
-
-        if (!document.hasFocus || document.hasFocus()) {
-          el.focus();
-        } else {
-          $el.trigger('focusin');
-        }
-      });
-    }
-  }
-}
-
-export function fireEvent(element, type, options = {}) {
-  if (!element) {
-    return;
-  }
-  let event;
-  if (KEYBOARD_EVENT_TYPES.indexOf(type) > -1) {
-    event = buildKeyboardEvent(type, options);
-  } else if (MOUSE_EVENT_TYPES.indexOf(type) > -1) {
-    let rect = element.getBoundingClientRect();
-    let x = rect.left + 1;
-    let y = rect.top + 1;
-    let simulatedCoordinates = {
-      screenX: x + 5,
-      screenY: y + 95,
-      clientX: x,
-      clientY: y
-    };
-    event = buildMouseEvent(type, jQuery.extend(simulatedCoordinates, options));
-  } else {
-    event = buildBasicEvent(type, options);
-  }
-  element.dispatchEvent(event);
-}
-
-function buildBasicEvent(type, options = {}) {
-  let event = document.createEvent('Events');
-  event.initEvent(type, true, true);
-  jQuery.extend(event, options);
-  return event;
-}
-
-function buildMouseEvent(type, options = {}) {
-  let event;
-  try {
-    event = document.createEvent('MouseEvents');
-    let eventOpts = jQuery.extend({}, DEFAULT_EVENT_OPTIONS, options);
-    event.initMouseEvent(
-      type,
-      eventOpts.canBubble,
-      eventOpts.cancelable,
-      window,
-      eventOpts.detail,
-      eventOpts.screenX,
-      eventOpts.screenY,
-      eventOpts.clientX,
-      eventOpts.clientY,
-      eventOpts.ctrlKey,
-      eventOpts.altKey,
-      eventOpts.shiftKey,
-      eventOpts.metaKey,
-      eventOpts.button,
-      eventOpts.relatedTarget);
-  } catch (e) {
-    event = buildBasicEvent(type, options);
-  }
-  return event;
-}
-
-function buildKeyboardEvent(type, options = {}) {
-  let event;
-  try {
-    event = document.createEvent('KeyEvents');
-    let eventOpts = jQuery.extend({}, DEFAULT_EVENT_OPTIONS, options);
-    event.initKeyEvent(
-      type,
-      eventOpts.canBubble,
-      eventOpts.cancelable,
-      window,
-      eventOpts.ctrlKey,
-      eventOpts.altKey,
-      eventOpts.shiftKey,
-      eventOpts.metaKey,
-      eventOpts.keyCode,
-      eventOpts.charCode
-    );
-  } catch (e) {
-    event = buildBasicEvent(type, options);
-  }
-  return event;
-}
